+++ tests/component/test_auth.py
@@ -5,7 +5,7 @@ import requests
 from ska_aaa_authhelpers import Role
 from ska_aaa_authhelpers.test_helpers import TEST_USER, mint_test_token
 
-from ..unit.util import SBDEFINITION_WITHOUT_ID_OR_METADATA_JSON
+from ..unit.util import VALID_PROJECT_WITHOUT_ID_JSON
 from . import ODT_URL, OSO_SERVICES_URL
 from .conftest import AUDIENCE
 
@@ -29,8 +29,8 @@ def test_user_extracted_from_auth():
         audience=AUDIENCE, scopes={"odt:readwrite"}, roles={Role.SW_ENGINEER}
     )
     response = requests.post(
-        f"{ODT_URL}/sbds",
-        data=SBDEFINITION_WITHOUT_ID_OR_METADATA_JSON,
+        f"{ODT_URL}/prjs",
+        data=VALID_PROJECT_WITHOUT_ID_JSON,
         headers={
             "Authorization": f"Bearer {token}",
             "Content-type": "application/json",