+++ tests/unit/odt/api/test_sbds.py
@@ -143,29 +143,6 @@ class TestSBDefinitionAPI:
         assert_json_is_equal(response.text, test_sbd.model_dump_json())
 
     @mock.patch("ska_oso_services.odt.api.sbds.validate_sbd")
-    def test_sbds_post_adds_status(self, mock_validate, client_with_uow_mock):
-        """
-        Check the sbds_post method also adds a status entity
-        """
-        mock_validate.return_value = {}
-        client, uow_mock = client_with_uow_mock
-        test_sbd = TestDataFactory.sbdefinition()
-        uow_mock.sbds.add.return_value = test_sbd
-        uow_mock.sbds.get.return_value = test_sbd
-
-        add_status_mock = mock.MagicMock()
-        uow_mock.sbds_status_history.add = add_status_mock
-
-        response = client.post(
-            f"{SBDS_API_URL}",
-            data=SBDEFINITION_WITHOUT_ID_JSON,
-            headers={"Content-type": "application/json"},
-        )
-
-        assert response.status_code == HTTPStatus.OK
-        add_status_mock.assert_called()
-
-    @mock.patch("ska_oso_services.odt.api.sbds.validate_sbd")
     def test_sbds_post_given_sbd_id_raises_error(self, mock_validate, client):
         """
         Check the sbds_post method returns a validation error if the user
@@ -248,30 +225,6 @@ class TestSBDefinitionAPI:
 
         assert response.status_code == HTTPStatus.OK
         assert_json_is_equal(response.text, test_sbd.model_dump_json())
-
-    @mock.patch("ska_oso_services.odt.api.sbds.validate_sbd")
-    def test_sbds_put_adds_status(self, mock_validate, client_with_uow_mock):
-        """
-        Check the sbds_put also adds a status entity
-        """
-        mock_validate.return_value = {}
-        client, uow_mock = client_with_uow_mock
-        uow_mock.sbds.__contains__.return_value = True
-        test_sbd = TestDataFactory.sbdefinition()
-        uow_mock.sbds.add.return_value = test_sbd
-        uow_mock.sbds.get.return_value = test_sbd
-
-        add_status_mock = mock.MagicMock()
-        uow_mock.sbds_status_history.add = add_status_mock
-
-        response = client.put(
-            f"{SBDS_API_URL}/sbd-mvp01-20200325-00001",
-            data=VALID_MID_SBDEFINITION_JSON,
-            headers={"Content-type": "application/json"},
-        )
-
-        assert response.status_code == HTTPStatus.OK
-        add_status_mock.assert_called()
 
     @mock.patch("ska_oso_services.odt.api.sbds.validate_sbd")
     def test_sbds_put_wrong_identifier(self, mock_validate, client):